+++ packages/orq-rc/src/models/operations/getagent.ts
@@ -946,7 +946,7 @@ export const GetAgentKnowledgeBases$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01K5QVJ9EJ7XANYHFG6QW4P43G"),
+  id: z.string().default("01K5R6RX51FMRPHRSWDK1BX78H"),
   knowledge_id: z.string(),
   configuration: z.union([
     z.lazy(() => GetAgentConfiguration2$inboundSchema),
@@ -973,7 +973,7 @@ export const GetAgentKnowledgeBases$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   GetAgentKnowledgeBases
 > = z.object({
-  id: z.string().default("01K5QVJ9EJ7XANYHFG6QW4P43G"),
+  id: z.string().default("01K5R6RX51FMRPHRSWDK1BX78H"),
   knowledgeId: z.string(),
   configuration: z.union([
     z.lazy(() => GetAgentConfiguration2$outboundSchema),